Add arg

+++ uicompositor/src/main/java/com/primogemstudio/advancedfmk/kui/yaml/Component.kt
@@ -1,5 +1,6 @@
 package com.primogemstudio.advancedfmk.kui.yaml
 
+import org.apache.logging.log4j.core.util.Integers
 import org.joml.Vector4f
 
 enum class ComponentType(val clz: String) {
@@ -46,6 +47,6 @@ class TextComponent(
 class GeometryLineComponent(
     var width: Float? = null,
     var color: List<Float>? = null,
-    var points: List<List<Float>>? = null,
+    var points: List<List<Int>>? = null,
     var filter: Map<String, String>? = null,
 ) : Component(ComponentType.GEOMETRY_LINE)
